requirements.txt
pylint
streamlit
pandas
requests
langchain-community
cohere
python-dotenv
transformers
torch 
torchvision 
torchaudio
langchain-cohere
